Added Date Sanitization, ListAll Funciton, and ListWhich function
Date Sanitization takes in many different date formats, and converts them to a 4 digit year string. ListAll function lists all the unique values for a given dimension (returns a list of genres, or list of production companies). ListWhich function returns all matches for a given value (ie if you want to know which movies had a budget of $300,000,000).